+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Added new CarPlay icons.
 * Added logic to automatically recenter puck after canceling preview mode.
 * Fixed Car Play route does not properly cancel when navigating solely from the Car Play application.
+* Added a new style for the puck and maneuver banner. 
 
 ## v3.5.0