+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 * Fixed Car Play route does not properly cancel when navigating solely from the Car Play application.
 * Added a new style for the puck and maneuver banner. 
 
+### Other
+
+* Fixed a resolved SDK name.
+
 ## v3.5.0
 
 ### Packaging